moving performance tests to a different case
